Adding quiz sample solutions
Sample solutions will just be one possibility for how you might have done things. It's not necessarily the most efficient approach, but the solutions are implemented such that they could have been done by a student with the current level of understanding on pace with course material. Send an email to me if you have any questions.